requirements.txt
gym==0.18.0
numpy==1.21.5
Pillow==9.1.1
scipy==1.4.1
stable-baselines3==1.3.0
statsmodels==0.11.1
torch==1.10.1
torchaudio==0.7.2
torchvision==0.11.2
